To develop an asp.net application that connects to and loads data from Oracle and MS SQL database
Objective:
To develop an asp.net application that connects to and loads data from
Oracle and MS SQL database

About the application:
This Application is basically programmed in .net to connect to Oracle
and MS SQL Server Database in order to fetch table data from selected
tables.
This application is credential protected which has LogIn Screen and only
limited Users can access application for security purpose.